update build and copying scripts

+++ tools/publish_wine_to_our_distro.sh
@@ -7,7 +7,14 @@
 load_mod alt
 
 PROJECTVERSION=$1
-[ -n "$PROJECTVERSION" ] || PROJECTVERSION=last
+if [ -z "$PROJECTVERSION" ] ; then
+	PROJECTVERSION=last
+	UNIPVERSION=testing
+	COMPONENT=unstable
+else
+	UNIPVERSION=stable
+	COMPONENT=nonfree
+fi
 
 [ -n "$BUILDNAME" ] || BUILDNAME=wine-etersoft
 
@@ -18,16 +25,21 @@ SOURCEPATH=$WINEPUB_PATH/$PROJECTVERSION/sources
 add_and_remove()
 {
 	local i=$2
+	[ -n "$i" ] || return
 	cd $1 || fatal "Can't cd"
 	#ls -l
 	local LIST="$i-[0-9]*.*.rpm"
+	for i in $LIST ; do
+	    [ -r "$i" ] || { echo "Skip $LIST (missed now)" ; return ; }
+	done
 	rm -fv $TP/$i-[0-9]*.rpm
+	#echo "## $(pwd) ## $LIST"
 	cp -flv $LIST $TP 2>/dev/null || cp -fv $LIST $TP || fatal "Can't copy $LIST"
 	cd - >/dev/null
 }
 
 # FROM TARGET
-copy_to()
+wine_copy_to()
 {
 	ARCH=$1
 	shift
@@ -35,7 +47,7 @@ copy_to()
 	[ "$ARCH" = "i586" ] || NARCH=$ARCH
 	local FPU="$WINEPUB_PATH/$PROJECTVERSION/WINE/$NARCH/$1"
 	local HASPFPU="$WINEPUB_PATH/$PROJECTVERSION/HASP/$NARCH/$1"
-	local TP="$2/$ARCH/RPMS.nonfree"
+	local TP="$2/$ARCH/RPMS.$COMPONENT"
 
 	add_and_remove "$FPU" wine-etersoft
 	add_and_remove "$FPU/extra" wine-etersoft-gl
@@ -53,18 +65,49 @@ copy_to()
 	#TP="$2/noarch/RPMS.addon"
 	add_and_remove "$FPU" etercifs
 
+}
+
+# FROM TARGET
+other_copy_to()
+{
+	ARCH=$1
+	shift
+	NARCH=
+	[ "$ARCH" = "i586" ] || NARCH=$ARCH
+	local TP="$2/$ARCH/RPMS.$COMPONENT"
+
+	FPU="/var/ftp/pub/Etersoft/RX@Etersoft/$UNIPVERSION/$NARCH/$1"
+	add_and_remove "$FPU" nx
+	add_and_remove "$FPU" nxclient
+	add_and_remove "$FPU" rx-etersoft
+
+	FPU="/var/ftp/pub/Etersoft/Postgre@Etersoft/$UNIPVERSION/$NARCH/$1"
+	for i in libpq5.2-9.0eter libpq5.2-9.0eter postgre-etersoft9.0 postgre-etersoft9.0 postgre-etersoft9.0-seltaaddon postgre-etersoft9.0-server; do
+		add_and_remove "$FPU" $i
+	done
+
 	set_binaryrepo $(basename $1)
 	ssh git.eter genbases -b $BINARYREPO
+}
+
+all_copy_to()
+{
+    wine_copy_to "$@"
+    other_copy_to "$@"
 }
 
 distro_path=/var/ftp/pub/Etersoft/LINUX@Etersoft
 
 arch=i586
-copy_to "$arch" ALTLinux/4.1 $distro_path/4.1/branch
-copy_to "$arch" ALTLinux/p5 $distro_path/p5/branch
-copy_to "$arch" ALTLinux/Sisyphus $distro_path/Sisyphus
-copy_to "$arch" ALTLinux/p6 $distro_path/p6/branch
+#copy_to "$arch" ALTLinux/4.1 $distro_path/4.1/branch
+all_copy_to "$arch" ALTLinux/p5 $distro_path/p5/branch
+all_copy_to "$arch" ALTLinux/Sisyphus $distro_path/Sisyphus
+all_copy_to "$arch" ALTLinux/p6 $distro_path/p6/branch
+all_copy_to "$arch" ALTLinux/p6 $distro_path/t6/branch
 
-#arch=x86_64
-#copy_to "$arch" ALTLinux/5.1 $distro_path/5.1/branch
-#copy_to "$arch" ALTLinux/Sisyphus $distro_path/Sisyphus
+arch=x86_64
+#copy_to "$arch" ALTLinux/4.1 $distro_path/4.1/branch
+other_copy_to "$arch" ALTLinux/p5 $distro_path/p5/branch
+other_copy_to "$arch" ALTLinux/Sisyphus $distro_path/Sisyphus
+other_copy_to "$arch" ALTLinux/p6 $distro_path/p6/branch
+other_copy_to "$arch" ALTLinux/p6 $distro_path/t6/branch
